+++ srcs/main/read_linetty.c
@@ -10,33 +10,12 @@
 ** コマンド実行中の標準入力への影響は未検証、
 */
 
-static bool	set_terminal_setting(void)
-{
-	struct termios	termios_p;
-	int				ret;
-
-	ret = tcgetattr(STDIN_FILENO, &termios_p);
-	if (ret == -1)
-	{
-		ft_putendl_fd(strerror(errno), 2);
-		return (false);
-	}
-	termios_p.c_lflag &= ~(ICANON | ECHO);
-	ret = tcsetattr(STDIN_FILENO, TCSANOW, &termios_p);
-	if (ret == -1)
-	{
-		ft_putendl_fd(strerror(errno), 2);
-		return (false);
-	}
-	return (true);
-}
-
-static char	*back_line(char *line, int *i)
+static void	back_line(char *line, int *i)
 {
 	line[*i - 1] = '\0';
 	*i = *i - 1;
 	//termcapを使ってカーソルを左へ移動、カーソル上の文字削除
-	return (line);
+	return ;
 }
 
 static char	*get_line(char *line, int i, int rc)
@@ -56,7 +35,7 @@ static char	*get_line(char *line, int i, int rc)
 		if (!ft_strncmp(c, UPKEY, 4) || !ft_strncmp(c, DOWNKEY, 4))
 			//termcapで現在のを消去、historyから表示。lineにコピー
 		if (rc != 0 && !ft_strncmp(c, "\177", 2))
-			line = back_line(line, &i);
+			back_line(line, &i);
 		if (rc == 1 && c[0] != '\n' && c[0] != '\177')
 		{
 			ft_putchar_fd(c[0], STDOUT_FILENO);
